+++ main/app/Resources/modules/components/main.jsx
@@ -1,4 +1,4 @@
-import React, {Fragment} from 'react'
+import React from 'react'
 import {PropTypes as T} from 'prop-types'
 import {Provider} from 'react-redux'
