+++ main/app/Resources/modules/data/types/enum-plus/components/option.jsx
@@ -1,5 +1,6 @@
 import React, {Component} from 'react'
 import {PropTypes as T} from 'prop-types'
+
 import {trans} from '#/main/app/intl/translation'
 
 class Option extends Component {
