+++ plugin/competency/Resources/modules/data/types/ability/index.js
@@ -1,19 +1,18 @@
 import {trans} from '#/main/app/intl/translation'
 
 import {AbilityDisplay} from '#/plugin/competency/data/types/ability/components/display'
-import {AbilityGroup} from '#/plugin/competency/data/types/ability/components/group'
+import {AbilityInput} from '#/plugin/competency/data/types/ability/components/input'
 
 const dataType = {
   name: 'ability',
   meta: {
-    creatable: false,
     icon: 'fa fa-fw fa fa-graduation-cap',
-    label: trans('ability', {}, 'competency'),
-    description: trans('ability_desc', {}, 'competency')
+    label: trans('ability', {}, 'data'),
+    description: trans('ability_desc', {}, 'data')
   },
   components: {
     details: AbilityDisplay,
-    form: AbilityGroup
+    form: AbilityInput
   }
 }
