+++ plugin/exo/Resources/modules/data/score-rules/components/group.jsx
@@ -6,7 +6,7 @@ import {FormGroup} from '#/main/app/content/form/components/group'
  * Overrides default form group to let each item in the
  * collection display its own errors.
  */
-const EnumGroup = props =>
+const ScoreRulesGroup = props =>
   <FormGroup
     {...props}
     error={typeof props.error === 'string' ? props.error : undefined}
@@ -14,8 +14,8 @@ const EnumGroup = props =>
     {props.children}
   </FormGroup>
 
-EnumGroup.propTypes = FormGroup.propTypes
+ScoreRulesGroup.propTypes = FormGroup.propTypes
 
 export {
-  EnumGroup
+  ScoreRulesGroup
 }
