+++ plugin/exo/Resources/modules/data/score-rules/index.js
@@ -1,7 +1,10 @@
 import isEmpty from 'lodash/isEmpty'
 import {chain, array, string, notBlank, lengthInRange} from '#/main/core/validation'
 
-import {ScoreRulesGroup} from '#/plugin/exo/data/score-rules/components/form-group'
+import {ScoreRulesGroup} from '#/plugin/exo/data/score-rules/components/group'
+import {ScoreRulesInput} from '#/plugin/exo/data/score-rules/components/input'
+
+// TODO : create score-rule and use `collection`
 
 const dataType = {
   name: 'score_rules',
@@ -46,7 +49,7 @@ const dataType = {
   }]),
 
   components: {
-    form: ScoreRulesGroup
+    input: ScoreRulesInput
   }
 }
