+++ plugin/planned-notification/Resources/modules/data/types/roles/prop-types.js
