+++ plugin/planned-notification/Resources/modules/plugin.js
@@ -8,8 +8,8 @@ import {registry} from '#/main/app/plugins/registry'
 registry.add('ClarolinePlannedNotificationBundle', {
   data: {
     types: {
-      'workspace_roles'       : () => { return import(/* webpackChunkName: "planned-notification-data-workspace_roles" */ '#/plugin/planned-notification/data/roles') },
-      'message'               : () => { return import(/* webpackChunkName: "planned-notification-data-message" */         '#/plugin/planned-notification/data/message') },
+      'workspace_roles'       : () => { return import(/* webpackChunkName: "planned-notification-data-workspace_roles" */ '#/plugin/planned-notification/data/types/roles') },
+      'message'               : () => { return import(/* webpackChunkName: "planned-notification-data-message" */         '#/plugin/planned-notification/data/types/message') },
       'planned_notifications' : () => { return import(/* webpackChunkName: "planned-notification-data-notifications" */   '#/plugin/planned-notification/data/types/notifications') }
     }
   }
